+++ CONTEXT.md
@@ -1,14 +1,13 @@
 # Context
 
-**Goal:** Execute the `bitvid-const-refactor-agent` task (daily scheduler).
-**Scope:** Identify and refactor duplicated numeric constants in `js/` to use canonical named constants.
-**Focus:** Refactor `5000` (timeout) to `SHORT_TIMEOUT_MS` and `60000` (timeout) to `LONG_TIMEOUT_MS` in selected files.
+**Goal:** Execute the `bitvid-decompose-agent` task (daily scheduler).
+**Scope:** Decompose `js/ui/profileModalController.js` by extracting hashtag preference logic into `js/ui/profileModal/ProfileHashtagController.js`.
+**Focus:** Reduce file size by moving cohesive hashtag-related methods and state.
 
 **Definition of Done:**
-- [ ] Identify candidate duplicates.
-- [ ] Refactor `js/ui/engagementController.js` (5000 -> SHORT_TIMEOUT_MS).
-- [ ] Refactor `js/ui/applicationBootstrap.js` (5000 -> SHORT_TIMEOUT_MS).
-- [ ] Refactor `js/webtorrent.js` (60000 -> LONG_TIMEOUT_MS).
+- [ ] Create `js/ui/profileModal/ProfileHashtagController.js`.
+- [ ] Move methods: `normalizeHashtagTag`, `formatHashtagTag`, `getResolvedHashtagPreferences`, `setHashtagStatus`, `refreshHashtagBackgroundStatus`, `populateHashtagPreferences`, `renderHashtagList`, `sanitizeHashtagList`, `createHashtagListItem`, `persistHashtagPreferences`, `handleAddHashtagPreference`, `handleRemoveHashtagPreference`, `handleHashtagPreferencesChange`, `describeHashtagPreferencesError`.
+- [ ] Integrate into `js/ui/profileModalController.js`.
 - [ ] Verify changes with `npm run lint` and `npm run test:unit`.
-- [ ] Update documentation (`DECISIONS.md`, `TEST_LOG.md`).
+- [ ] Update `scripts/check-file-size.mjs` baseline.
 - [ ] Update `docs/agents/AGENT_TASK_LOG.csv`.
