+++ scripts/check-file-size.mjs
@@ -26,7 +26,7 @@ const GROWTH_MARGIN = 50; // lines — allowlisted files may grow by this much f
 // These files are known to be oversized and are decomposition targets.
 // They must NOT grow beyond (recorded + GROWTH_MARGIN).
 const GRANDFATHERED = {
-  "js/ui/profileModalController.js": 8753,
+  "js/ui/profileModalController.js": 8130,
   "js/ui/components/VideoModal.js": 6284,
   "js/app.js": 5602,
   "js/channelProfile.js": 5529,
